Moved the parser function to connector file
