+++ apps/netsuite/connector.py
@@ -1452,7 +1452,6 @@ class NetSuiteConnector:
             logger.info('Charge Card Error - %s', raw_response.text)
 
             try:
-                from apps.netsuite.helpers import parse_error_and_get_message
                 error_message = parse_error_and_get_message(raw_response.text)
             except Exception as e:
                 logger.info('Error while parsing error message - %s', e)
@@ -2079,3 +2078,38 @@ class NetSuiteConnector:
 
         created_vendor_payment = self.connection.vendor_payments.post(vendor_payment_payload)
         return created_vendor_payment
+
+
+def parse_error_and_get_message(raw_response):
+    try:
+        if raw_response == '<HTML><HEAD>' or raw_response == '<html>':
+            return 'HTML bad response from NetSuite'
+        raw_response = raw_response.replace("'", '"')\
+            .replace("False", 'false')\
+            .replace("True", 'true')\
+            .replace("None", 'null')
+        parsed_response = json.loads(raw_response)
+        return get_message_from_parsed_error(parsed_response)
+    except Exception:
+        raw_response = raw_response.replace('"creditCardCharge"', 'creditCardCharge')\
+            .replace('""{', '{').replace('}""', '}')\
+            .replace('"{', '{').replace('}"', '}')\
+            .replace('\\"', '"').replace('\\', '')\
+            .replace('"https://', "'https://").replace('.html"', ".html'")\
+            .replace('="', "=").replace('">', ">")
+        parsed_response = json.loads(raw_response)
+        return get_message_from_parsed_error(parsed_response)
+
+
+def get_message_from_parsed_error(parsed_response):
+    try:
+        if 'error' in parsed_response:
+            if 'message' in parsed_response['error']:
+                if 'message' in parsed_response['error']['message']:
+                    return parsed_response['error']['message']['message']
+                return parsed_response['error']['message']
+        elif 'message' in parsed_response:
+            if 'message' in parsed_response['message']:
+                return parsed_response['message']['message']
+    except Exception:
+        raise
